UI 적용 완료, 이펙트 애니메이션 구현중

+++ Project_JSS/Assets/Project/Scripts/Managers/SaveManager.cs
@@ -7,7 +7,7 @@ public class SaveData
 {
     public List<ItemData> items = new List<ItemData>();
     public List<GeneratorSO> generators = new List<GeneratorSO>();
-    public float energy;
+    public int energy;
     public int score;
     public string playerName;
